refactor(login): split into server page + client form with mailEnabled prop
